requirements.txt
#
# The reference packages of NexLib
# The minimum Python Version is 3.9; prefer 3.11.5+
#
asyncua>=1.0.4
backoff>=2.2.1
loguru>=0.7.2
mariadb>=1.1.11
matplotlib>=3.9.1
numpy>=1.26.4
orjson>=3.9.7
paho-mqtt>=2.1.0
pandas>=2.2.2
pika>=1.3.2
pymongo>=4.5.0
pyYAML>=6.0.1
regex>=2024.5.15
requests>=2.32.3
requests-file>=2.1.0
rsa>=4.9
websockets>=11.0.3
#
#
# For AI related stuff
#
#haystack-ai>=2.2.4
#huggingface-hub>=0.23.4
#onnx>=1.16.1
#openai>=1.42.0
#openvino>=2024.2.0